+++ Libraries/LibELF/DynamicLoader.h
@@ -75,9 +75,6 @@ public:
     template<typename F>
     void for_each_needed_library(F) const;
 
-    DynamicObject::SymbolLookupFunction m_global_symbol_lookup_func { nullptr };
-    void set_global_symbol_lookup_function(DynamicObject::SymbolLookupFunction func) { m_global_symbol_lookup_func = func; }
-
     VirtualAddress text_segment_load_address() const { return m_text_segment_load_address; }
     bool is_dynamic() const { return m_elf_image.is_dynamic(); }
